refactor(#792): split infra-yaml lib_tests into base + roundtrip + misc suites (sweep)
lib_tests.rs 1597 -> 561, + lib_roundtrip_tests.rs 531 + lib_misc_tests.rs 531 (all
under cap; off debt). Three-way themed split: base keeps the fixtures (core_block/
first_model/first_model_optional become pub(super)), roundtrip holds the serialize/
edge-case tests, misc holds device/NAM/multi-chain/error tests — both import the
fixtures via super::tests. infra-yaml tests pass, zero warnings.
